CODE RUB: Move Exceptions To Model Folder
