+++ NajotBooking.Api/Models/Orders/Exceptions/OrderValidationException.cs
@@ -5,11 +5,12 @@
 
 using Xeptions;
 
-namespace NajotBooking.Api.Services.Foundations.Orders.Exceptions
+namespace NajotBooking.Api.Models.Orders.Exceptions
 {
-    public class NullOrderException : Xeption
+    public class OrderValidationException : Xeption
     {
-        public NullOrderException() : base(message: "Order is null.")
+        public OrderValidationException(Xeption innerException)
+            : base(message: "Order error occured, fix the errors and try again.", innerException)
         { }
     }
 }
